Basic execution of queued nodes
Again, this is *not* yet applicable to arbitrary networks, but the basic
principle seems to work. The pump is now capable of deciding whether any
edge is ready for processing. While processing certain edges, the status
of all nodes is monitored. Nodes that are ready for execution are queued
and will subsequently be executed.