Implemented node conversion to Makefile rules
So far, this is not really useful because the rule is unable to process
the output files correctly---it lacks knowledge about all of the output
files because an individual node is unable to access any dependent node
in the workflow for reasons of encapsulation.

This is where I need a separate function in the main "Pump" class.